+++ nifi-analyticZooModelServing-processor/src/main/java/com/ssqcyy/nifi/processor/UserItemPair.java
@@ -5,20 +5,20 @@ package com.ssqcyy.nifi.processor;
  */
 public class UserItemPair {
 
-    private final Integer userId;
-    private final Integer itemId;
+    private final Float userId;
+    private final Float itemId;
 
-    public UserItemPair(Integer userId, Integer itemId){
+    public UserItemPair(Float userId, Float itemId){
 
         this.userId = userId;
         this.itemId = itemId;
     }
 
-    public Integer getUserId() {
+    public Float getUserId() {
         return userId;
     }
 
-    public Integer getItemId() {
+    public Float getItemId() {
         return itemId;
     }
 }
